+++ capabilities/base/files/src/{{Name}}.Infrastructure/DependencyInjection.cs
@@ -7,7 +7,7 @@ public static class DependencyInjection
 {
     public static IServiceCollection AddInfrastructureServices(this IServiceCollection services, IConfiguration config)
     {
-        // devstart:infrastructure-services
+        // dev-start-dotnet:infrastructure-services
         return services;
     }
 }